feat: refactored all new semesterApiService.ts file
