--- .vscode/settings.json
@@ -1,6 +1,0 @@
-{
-  "workbench.colorCustomizations": {
-    "minimap.background": "#00000000",
-    "scrollbar.shadow": "#00000000"
-  }
-}
